requirements.txt
python-dotenv
langchain_groq 
matplotlib
streamlit
tabulate
groq
langchain_experimental
pandas
